# swagger.yaml
openapi: 3.0.3
info:
  title: Profile Builder API
  description: API documentation for Creating a Profile Builder Application with Admin Control and some advanced features.s
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/profiles/{profile_id}:
    delete:
      summary: Delete User Profile
      tags:
        - Profiles
      description: Delete a profile using specific profile ID.
      operationId: deleteProfile
      parameters:
        - name: profile_id
          in: path
          description: The ID of the profile to be deleted.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile deletion response.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Profile deleted successfully.
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Resource not found for the given profile ID.
        "401":
          description: Unauthorized user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    example: 401
                  error_message:
                    type: string
                    example: invalid token
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    example: 500
                  error_message:
                    type: string
                    example: internal server error
        "502":
          description: Invalid request data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: integer
                    example: 502
                  error_message:
                    type: string
                    example: invalid request data
...